+++ test/models/dictionary.rb
@@ -1,0 +1,2 @@
+class Dictionary < Book
+end