+++ test/schema/generic_schema.rb
@@ -64,6 +64,7 @@ ActiveRecord::Schema.define do
     t.integer :topic_id
     t.boolean :for_sale, default: true
     t.integer :status, default: 0
+    t.string :type
   end
 
   create_table :chapters, force: :cascade do |t|
@@ -73,7 +74,7 @@ ActiveRecord::Schema.define do
     t.datetime :updated_at
   end
 
-  create_table :end_notes, force: :cascade do |t|
+  create_table :end_notes, primary_key: :end_note_id, force: :cascade do |t|
     t.string :note
     t.integer :book_id, null: false
     t.datetime :created_at
@@ -133,12 +134,21 @@ ActiveRecord::Schema.define do
     t.string :discountable_type
   end
 
-  create_table :rules, force: :cascade do |t|
+  create_table :rules, id: false, force: :cascade do |t|
+    t.integer :id
     t.string :condition_text
     t.integer :question_id
   end
 
   create_table :questions, force: :cascade do |t|
     t.string :body
+  end
+
+  create_table :vendors, force: :cascade do |t|
+    t.string :name, null: true
+    t.text :preferences
+    t.text :data
+    t.text :config
+    t.text :settings
   end
 end