+++ test/sqlite3/import_test.rb
@@ -1,6 +1,7 @@
 require File.expand_path(File.dirname(__FILE__) + '/../test_helper')
 
 should_support_recursive_import
+should_support_on_duplicate_key_ignore
 
 describe "#supports_imports?" do
   context "and SQLite is 3.7.11 or higher" do
@@ -47,6 +48,20 @@ describe "#import" do
       result = Topic.import Build(2500, :topics)
       assert_equal 5, result.num_inserts, "Failed to issue a two INSERT statements. Make sure you have a supported version of SQLite3 (3.7.11 or higher) installed"
       assert_equal 2500, Topic.count, "Failed to insert all records. Make sure you have a supported version of SQLite3 (3.7.11 or higher) installed"
+    end
+  end
+
+  context "with :on_duplicate_key_update" do
+    let(:topics) { Build(1, :topics) }
+
+    it "should log a warning message" do
+      log = StringIO.new
+      logger = Logger.new(log)
+      logger.level = Logger::WARN
+      ActiveRecord::Base.connection.stubs(:logger).returns(logger)
+
+      Topic.import topics, on_duplicate_key_update: true
+      assert_match(/Ignoring on_duplicate_key_update/, log.string)
     end
   end
 end