+++ test/support/mysql/import_examples.rb
@@ -4,6 +4,7 @@ def should_support_mysql_import_functionality
   ActiveRecord::Base.connection.execute "set sql_mode='STRICT_ALL_TABLES'"
 
   should_support_basic_on_duplicate_key_update
+  should_support_on_duplicate_key_ignore
 
   describe "#import" do
     context "with :on_duplicate_key_update and validation checks turned off" do